requirements.txt
Flask==3.1.0
python-dotenv==1.0.1
Requests==2.32.3
twilio==9.3.7
gunicorn==20.1.0
